+++ miles/ray/actor_group.py
@@ -81,7 +81,7 @@ class RayTrainGroup:
             actor_impl = MegatronTrainRayActor
 
         else:
-            from miles.backends.fsdp_utils import FSDPTrainRayActor
+            from miles.backends.experimental.fsdp_utils import FSDPTrainRayActor
 
             actor_impl = FSDPTrainRayActor
